avanzo init service y dao

+++ examen/src/main/java/com/cash/examen/dao/UserDAO.java
@@ -9,7 +9,7 @@ public interface UserDAO {
 
     List<User> getAllUsers();
 
-    User getUser(int user);
+    User getUser(long user);
 
     void addUser(User user);
